feat: statistic refactor
